Added Part version, geometry type 1. Source code splitted and re-arranged. A lot of background cleaning a modifications for easier adding new types and versions.
